Switch modal workflow JS to onload to support Wagtail>=2.2
